pyproject.toml
[project]
name = "video-materialize"
version = "0.1.0"
description = "Video materialization policy experiments for repeated-query video QA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.70.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "tqdm>=4.67.3",
    "yt-dlp>=2026.3.17",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vm"]
